Convert the the USB device name UTF8
The USB device name is a UTF16 string but the `spndev.name` is a `char *`.
This causes confusion when someone tries to `printf` the device name and
gets only the first character, because the second byte is 0.

To avoid this convert the device name to UTF8. For devices whose names are
ASCII (most cases) this will result in a plain ASCIIZ string and avoid
confusion.

At the same time the driver API remains "simple" for embedded platforms
where the USB HID implementation will not be HIDAPI and Unicode support
may be missing (or too large to compile in).

Fixes #6