Serial (RS-232) device support improved
Fixed a bug (crash in debug mode) with Spaceball 4000 FLX with
Fw: "Firmware version 2.42 created on 24-Oct-1997."

Initially it is detected as a 2003c - and this is normal, because both
use the 2.42 firmware version.

The first time a button above 8 (9, A, B, C written on the button) was
pressed the driver would access unallocated memory because `dev->bn_name`
only had space for 8 buttons. The `dev` structure is now properly
reinitialized.

Tested with:
Spaceball 4000 FLX with Fw: "Firmware version 2.42 created on 24-Oct-1997."