Fix the number of buttons and names for "Magellan SpaceMouse Plus XT USB"
There are 11 buttons and the "*" is the 8th (counting from zero) button.

I called the two buttons on the side of the "hat" "L" (9) and "R" (10) even
though I called them "+" and "-" respectively for the serial version
of the "Magellan SpaceMouse" in serdev.c - see e17b5289d6. I need to decide
on one or the other in the future.

Fixed the "Opened USB device" message by copying the cinfo->product_string
wchar_t * passed to wcsrtombs, because it sets it to NULL.

Tested with:
    Magellan SpaceMouse Plus XT USB